Address review nits in NaN-safe progress fix
Agent-Logs-Url: https://github.com/M1314/FCD/sessions/bd5d367e-2d2b-4d89-abd7-0195f65aae94

+++ lib/src/features/courses/presentation/course_player_page.dart
@@ -177,7 +177,7 @@ class _CoursePlayerPageState extends State<CoursePlayerPage>
     if (!raw.isFinite) {
       return 0;
     }
-    return raw.clamp(0.0, 1.0).toDouble();
+    return raw.clamp(0.0, 1.0);
   }
 
   @override
@@ -572,7 +572,7 @@ class _CoursePlayerPageState extends State<CoursePlayerPage>
             return;
           }
           setState(() {
-            _downloadProgress = raw.clamp(0.0, 1.0).toDouble();
+            _downloadProgress = raw.clamp(0.0, 1.0);
           });
         },
       );
